move useLibp2p into its own folder
